+++ mailer/settings.py
@@ -1,5 +1,6 @@
 from typing import Optional, Set
-from pydantic import BaseSettings, EmailStr, AnyHttpUrl
+from pydantic import BaseSettings, EmailStr, AnyHttpUrl, validator
+from pgpy import PGPKey
 
 from . import __about__
 
@@ -21,8 +22,27 @@ class Settings(BaseSettings):
     smtp_user: str
     smtp_password: str
 
+    pgp_public_key: Optional[PGPKey] = None
+
     cors_origins: Set[AnyHttpUrl] = set()
 
     recaptcha_secret_key: Optional[str]
 
     sentry_dsn: Optional[str] = None
+
+    @validator("pgp_public_key", pre=True)
+    def validate_pgp_public_key(cls, v: Optional[str]) -> Optional[PGPKey]:
+        from pgpy.errors import PGPError
+
+        if v:
+            try:
+                key, _ = PGPKey.from_blob(v.encode("utf-8"))
+            except (ValueError, PGPError):
+                raise ValueError("Invalid PGP public key: cannot load the key")
+
+            if not key.is_public:
+                raise ValueError("Invalid PGP public key: key is private")
+
+            return key
+
+        return None