+++ tests/test_api.py
@@ -2,15 +2,65 @@ import pytest
 
 from http import HTTPStatus
 
+from . import utils
 
-valid_origin = "https://domain.me"
-valid_recaptcha_secret = "valid-recaptcha-secret"
-valid_recaptcha_response = "valid-recaptcha-response"
+
+# ------------------------------------------------------------------------------
 
 
 @pytest.fixture(scope="function")
-def enable_cors_origins(monkeypatch):
-    monkeypatch.setenv("CORS_ORIGINS", f'["{valid_origin}"]')
+def enable_cors_origins(monkeypatch, faker):
+    origin = faker.url()
+    monkeypatch.setenv("CORS_ORIGINS", f'["{origin}"]')
+    return origin
+
+
+# ------------------------------------------------------------------------------
+
+
+@pytest.fixture(scope="function")
+def mock_smtp(mocker):
+    mock_client = mocker.patch("smtplib.SMTP", autospec=True)
+
+    return mock_client
+
+
+@pytest.fixture(scope="function")
+def mock_smtp_connect_error(mock_smtp):
+    from smtplib import SMTPConnectError
+
+    mock_smtp.side_effect = SMTPConnectError(400, "error")
+
+    return mock_smtp
+
+
+@pytest.fixture(scope="function")
+def mock_smtp_auth_error(mock_smtp, mocker):
+    from smtplib import SMTPAuthenticationError
+
+    mock_smtp.return_value.login = mocker.Mock(
+        side_effect=SMTPAuthenticationError(401, "error")
+    )
+
+    return mock_smtp
+
+
+@pytest.fixture(scope="function")
+def mock_smtp_send_error(mock_smtp, mocker):
+    from smtplib import SMTPDataError
+
+    mock_smtp.return_value.send_message = mocker.Mock(
+        side_effect=SMTPDataError(402, "error")
+    )
+
+    return mock_smtp
+
+
+# ------------------------------------------------------------------------------
+
+
+valid_recaptcha_secret = "valid-recaptcha-secret"
+valid_recaptcha_response = "valid-recaptcha-response"
 
 
 @pytest.fixture(scope="function")
@@ -21,28 +71,6 @@ def enable_recaptcha(monkeypatch):
 @pytest.fixture(scope="function")
 def enable_recaptcha_invalid_secret(monkeypatch, faker):
     monkeypatch.setenv("RECAPTCHA_SECRET_KEY", faker.pystr())
-
-
-@pytest.fixture(scope="function")
-def mock_smtp(mocker):
-    mock_client = mocker.patch("emails.backend.SMTPBackend.sendmail", autospec=True)
-    mock_client.return_value.status_code = None
-
-    return mock_client
-
-
-@pytest.fixture(scope="function")
-def mock_smtp_success(mock_smtp):
-    mock_smtp.return_value.status_code = 250
-
-    return mock_smtp
-
-
-@pytest.fixture(scope="function")
-def mock_smtp_unavailable(mock_smtp):
-    mock_smtp.return_value.status_code = 421
-
-    return mock_smtp
 
 
 @pytest.fixture(scope="function")
@@ -89,6 +117,19 @@ def mock_recaptcha_verify_api(responses, faker):
     return responses
 
 
+# ------------------------------------------------------------------------------
+
+
+@pytest.fixture(scope="function")
+def enable_pgp_public_key(monkeypatch, faker):
+    pgp_key = utils.generate_pgp_key_pair(faker.name(), faker.email())
+    monkeypatch.setenv("PGP_PUBLIC_KEY", str(pgp_key.pubkey))
+    return pgp_key
+
+
+# ------------------------------------------------------------------------------
+
+
 @pytest.fixture(scope="function")
 def params_success(faker):
     return {
@@ -118,7 +159,7 @@ def test_get_api_info_success(app_client):
 # ------------------------------------------------------------------------------
 
 
-def test_send_mail_success(app_client, mock_smtp_success, params_success):
+def test_send_mail_success(app_client, mock_smtp, params_success):
     params = params_success
 
     response = app_client.post("/api/mail", json=params)
@@ -131,29 +172,70 @@ def test_send_mail_success(app_client, mock_smtp_success, params_success):
     assert data["message"] == params["message"]
     assert data["honeypot"] == params["honeypot"]
 
+    assert mock_smtp.call_count == 1
+    assert mock_smtp.return_value.login.call_count == 1
+    assert mock_smtp.return_value.send_message.call_count == 1
+    assert mock_smtp.return_value.quit.call_count == 1
 
-def test_send_mail_smtp_unavailable(app_client, mock_smtp_unavailable, params_success):
+    sent_msg = mock_smtp.return_value.send_message.call_args.args[0].as_string()
+    utils.assert_plain_email(
+        sent_msg, params["email"], params["name"], params["subject"], params["message"]
+    )
+
+
+def test_send_mail_smtp_connect_failed(
+    app_client, mock_smtp_connect_error, params_success
+):
     params = params_success
 
     response = app_client.post("/api/mail", json=params)
     assert response.status_code == HTTPStatus.UNAUTHORIZED
 
+    assert mock_smtp_connect_error.call_count == 1
+    assert mock_smtp_connect_error.return_value.login.call_count == 0
+    assert mock_smtp_connect_error.return_value.send_message.call_count == 0
+    assert mock_smtp_connect_error.return_value.quit.call_count == 0
 
-def test_send_mail_none(app_client, mock_smtp_success):
+
+def test_send_mail_smtp_login_failed(app_client, mock_smtp_auth_error, params_success):
+    params = params_success
+
+    response = app_client.post("/api/mail", json=params)
+    assert response.status_code == HTTPStatus.UNAUTHORIZED
+
+    assert mock_smtp_auth_error.call_count == 1
+    assert mock_smtp_auth_error.return_value.login.call_count == 1
+    assert mock_smtp_auth_error.return_value.send_message.call_count == 0
+    assert mock_smtp_auth_error.return_value.quit.call_count == 0
+
+
+def test_send_mail_smtp_send_failed(app_client, mock_smtp_send_error, params_success):
+    params = params_success
+
+    response = app_client.post("/api/mail", json=params)
+    assert response.status_code == HTTPStatus.UNAUTHORIZED
+
+    assert mock_smtp_send_error.call_count == 1
+    assert mock_smtp_send_error.return_value.login.call_count == 1
+    assert mock_smtp_send_error.return_value.send_message.call_count == 1
+    assert mock_smtp_send_error.return_value.quit.call_count == 0
+
+
+def test_send_mail_none(app_client, mock_smtp):
     params = {}
 
     response = app_client.post("/api/mail", json=params)
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_empty_fields(app_client, mock_smtp_success):
+def test_send_mail_empty_fields(app_client, mock_smtp):
     params = {"email": "", "name": "", "subject": "", "message": "", "honeypot": ""}
 
     response = app_client.post("/api/mail", json=params)
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_empty_email(app_client, mock_smtp_success, params_success):
+def test_send_mail_empty_email(app_client, mock_smtp, params_success):
     params = params_success
     params["email"] = ""
 
@@ -161,7 +243,7 @@ def test_send_mail_empty_email(app_client, mock_smtp_success, params_success):
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_empty_name(app_client, mock_smtp_success, params_success):
+def test_send_mail_empty_name(app_client, mock_smtp, params_success):
     params = params_success
     params["name"] = ""
 
@@ -169,7 +251,7 @@ def test_send_mail_empty_name(app_client, mock_smtp_success, params_success):
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_empty_subject(app_client, mock_smtp_success, params_success):
+def test_send_mail_empty_subject(app_client, mock_smtp, params_success):
     params = params_success
     params["subject"] = ""
 
@@ -177,7 +259,7 @@ def test_send_mail_empty_subject(app_client, mock_smtp_success, params_success):
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_empty_message(app_client, mock_smtp_success, params_success):
+def test_send_mail_empty_message(app_client, mock_smtp, params_success):
     params = params_success
     params["message"] = ""
 
@@ -185,7 +267,7 @@ def test_send_mail_empty_message(app_client, mock_smtp_success, params_success):
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_bad_email(app_client, mock_smtp_success, params_success):
+def test_send_mail_bad_email(app_client, mock_smtp, params_success):
     params = params_success
     params["email"] = "joe@doe"
 
@@ -193,7 +275,7 @@ def test_send_mail_bad_email(app_client, mock_smtp_success, params_success):
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_too_long_name(app_client, mock_smtp_success, params_success, faker):
+def test_send_mail_too_long_name(app_client, mock_smtp, params_success, faker):
     params = params_success
     params["name"] = faker.text(max_nb_chars=100)
 
@@ -201,9 +283,7 @@ def test_send_mail_too_long_name(app_client, mock_smtp_success, params_success, 
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_too_long_subject(
-    app_client, mock_smtp_success, params_success, faker
-):
+def test_send_mail_too_long_subject(app_client, mock_smtp, params_success, faker):
     params = params_success
     params["subject"] = faker.text(max_nb_chars=2000)
 
@@ -211,9 +291,7 @@ def test_send_mail_too_long_subject(
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_too_long_message(
-    app_client, mock_smtp_success, params_success, faker
-):
+def test_send_mail_too_long_message(app_client, mock_smtp, params_success, faker):
     params = params_success
     params["message"] = faker.text(max_nb_chars=2000)
 
@@ -221,9 +299,7 @@ def test_send_mail_too_long_message(
     assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
 
 
-def test_send_mail_non_empty_honeypot(
-    app_client, mock_smtp_success, params_success, faker
-):
+def test_send_mail_non_empty_honeypot(app_client, mock_smtp, params_success, faker):
     params = params_success
     params["honeypot"] = faker.text()
 
@@ -232,18 +308,18 @@ def test_send_mail_non_empty_honeypot(
 
 
 def test_send_mail_matching_cors_origin(
-    enable_cors_origins, app_client, mock_smtp_success, params_success
+    enable_cors_origins, app_client, mock_smtp, params_success
 ):
     params = params_success
 
-    headers = {"Origin": valid_origin}
+    headers = {"Origin": enable_cors_origins}
 
     response = app_client.post("/api/mail", json=params, headers=headers)
     assert response.status_code == HTTPStatus.OK
 
 
 def test_send_mail_unmatched_cors_origin(
-    enable_cors_origins, app_client, mock_smtp_success, params_success, faker
+    enable_cors_origins, app_client, mock_smtp, params_success, faker
 ):
     params = params_success
 
@@ -254,11 +330,7 @@ def test_send_mail_unmatched_cors_origin(
 
 
 def test_send_mail_recaptcha_success(
-    enable_recaptcha,
-    app_client,
-    mock_smtp_success,
-    mock_recaptcha_verify_api,
-    params_success,
+    enable_recaptcha, app_client, mock_smtp, mock_recaptcha_verify_api, params_success
 ):
     params = params_success
     params["g-recaptcha-response"] = valid_recaptcha_response
@@ -270,7 +342,7 @@ def test_send_mail_recaptcha_success(
 def test_send_mail_recaptcha_invalid_secret(
     enable_recaptcha_invalid_secret,
     app_client,
-    mock_smtp_success,
+    mock_smtp,
     mock_recaptcha_verify_api,
     params_success,
 ):
@@ -282,11 +354,7 @@ def test_send_mail_recaptcha_invalid_secret(
 
 
 def test_send_mail_recaptcha_no_response(
-    enable_recaptcha,
-    app_client,
-    mock_smtp_success,
-    mock_recaptcha_verify_api,
-    params_success,
+    enable_recaptcha, app_client, mock_smtp, mock_recaptcha_verify_api, params_success
 ):
     params = params_success
     params["g-recaptcha-response"] = ""
@@ -298,7 +366,7 @@ def test_send_mail_recaptcha_no_response(
 def test_send_mail_recaptcha_invalid_response(
     enable_recaptcha,
     app_client,
-    mock_smtp_success,
+    mock_smtp,
     mock_recaptcha_verify_api,
     params_success,
     faker,
@@ -308,3 +376,105 @@ def test_send_mail_recaptcha_invalid_response(
 
     response = app_client.post("/api/mail", json=params)
     assert response.status_code == HTTPStatus.UNAUTHORIZED
+
+
+def test_send_pgp_mail_success(
+    enable_pgp_public_key, app_client, mock_smtp, params_success
+):
+    params = params_success
+
+    response = app_client.post("/api/mail", json=params)
+    assert response.status_code == HTTPStatus.OK
+
+    data = response.json()
+    assert data["email"] == params["email"]
+    assert data["name"] == params["name"]
+    assert data["subject"] == params["subject"]
+    assert data["message"] == params["message"]
+    assert data["honeypot"] == params["honeypot"]
+
+    assert mock_smtp.call_count == 1
+    assert mock_smtp.return_value.login.call_count == 1
+    assert mock_smtp.return_value.send_message.call_count == 1
+    assert mock_smtp.return_value.quit.call_count == 1
+
+    message = mock_smtp.return_value.send_message.call_args.args[0]
+    sent_msg = message.as_string()
+    embedded_pub_key = utils.assert_pgp_email(
+        sent_msg,
+        params["email"],
+        params["name"],
+        params["subject"],
+        params["message"],
+        enable_pgp_public_key,
+        None,
+    )
+    assert embedded_pub_key is None
+
+
+def test_send_pgp_mail_with_attached_public_key_success(
+    enable_pgp_public_key, app_client, mock_smtp, params_success, faker
+):
+    sender_key = utils.generate_pgp_key_pair(faker.name(), faker.email())
+
+    params = params_success
+    params["public_key"] = str(sender_key.pubkey)
+
+    response = app_client.post("/api/mail", json=params)
+    assert response.status_code == HTTPStatus.OK
+
+    data = response.json()
+    assert data["email"] == params["email"]
+    assert data["name"] == params["name"]
+    assert data["subject"] == params["subject"]
+    assert data["message"] == params["message"]
+    assert data["honeypot"] == params["honeypot"]
+
+    assert mock_smtp.call_count == 1
+    assert mock_smtp.return_value.login.call_count == 1
+    assert mock_smtp.return_value.send_message.call_count == 1
+    assert mock_smtp.return_value.quit.call_count == 1
+
+    message = mock_smtp.return_value.send_message.call_args.args[0]
+    sent_msg = message.as_string()
+    embedded_pub_key = utils.assert_pgp_email(
+        sent_msg,
+        params["email"],
+        params["name"],
+        params["subject"],
+        params["message"],
+        enable_pgp_public_key,
+        sender_key.pubkey,
+    )
+    assert embedded_pub_key == str(sender_key.pubkey)
+
+    email_response = faker.text()
+    pgp_response = utils.encrypt_pgp_message(embedded_pub_key, email_response)
+    plain_response = utils.decrypt_pgp_message(str(sender_key), pgp_response)
+    assert plain_response == email_response
+
+
+def test_send_pgp_mail_with_attached_public_key_private(
+    enable_pgp_public_key, app_client, mock_smtp, params_success, faker
+):
+    sender_key = utils.generate_pgp_key_pair(faker.name(), faker.email())
+
+    params = params_success
+    params["public_key"] = str(sender_key)
+
+    response = app_client.post("/api/mail", json=params)
+    assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY
+
+
+def test_send_pgp_mail_with_attached_public_key_invalid(
+    enable_pgp_public_key, app_client, mock_smtp, params_success, faker
+):
+    from base64 import urlsafe_b64encode
+
+    pgp_key = urlsafe_b64encode(faker.binary()).decode("utf-8")
+
+    params = params_success
+    params["public_key"] = pgp_key
+
+    response = app_client.post("/api/mail", json=params)
+    assert response.status_code == HTTPStatus.UNPROCESSABLE_ENTITY